chore(metricprovider): hack dataFetcher to do the meta transform in consuming app
